+++ client/test/fixture/index.vue
@@ -8,10 +8,18 @@
     .test-vue-scoped-class-css{
         color: red;
     }
+
+    .test-vue-class-keyed{
+        color: red;
+    }
+
+    .test_vue_class_keyed{
+        color: red;
+    }
 </style>
 
 <style scoped lang="scss">
-    .test-vue-scoped-class-scss{
+    .test-vue-scoped-class{
         &-scss{
             color: red;
         }
@@ -19,7 +27,7 @@
 </style>
 
 <style scoped lang="less">
-    .test-vue-scoped-class-less{
+    .test-vue-scoped-class{
         &-less{
             color: red;
         }
@@ -30,17 +38,18 @@
     <div class="test-vue-scoped-class">Test Text</div>
     <div class="test-vue-scoped-class-css">Test Text</div>
     <div class="test-vue-scoped-class-scss">Test Text</div>
-    <div class="test-vue-scoped-class-less">Test Text</div>
+    <div :class="'test-vue-scoped-class-less'">Test Text</div>
 
     <div class="test-vue-import-class">Test Text</div>
     <div class="class-style-imported">Test Text</div>
+
+    <div :class="{'test-vue-class-keyed': boolean}"></div>
+    <div :class="{test_vue_class_keyed: boolean}"></div>
+    <div :class="vue_variable_not_class"></div>
 </template>
 
 
 <script>
-
-// Import in a <script> is not working recently.
 import "./css/test.scss"
 import "./node_modules/test-module/style-imported.scss"
-
 </script>