+++ server/src/languages/language-ids/language-ids.ts
@@ -1,11 +1,4 @@
 export namespace LanguageIds {
-	export function isHTMLSyntax(languageId: AllLanguageId): boolean {
-		return languageId === 'html'
-			|| languageId === 'jsx'
-			|| languageId === 'tsx'
-			|| languageId === 'js'
-			|| languageId === 'ts'
-	}
 
 	export function isCSSSyntax(languageId: AllLanguageId): boolean {
 		return languageId === 'css'
@@ -21,7 +14,16 @@ export namespace LanguageIds {
 	}
 
 	export function isScriptSyntax(languageId: AllLanguageId): boolean {
-		return isHTMLSyntax(languageId) && languageId !== 'html'
+		return languageId === 'jsx'
+			|| languageId === 'tsx'
+			|| languageId === 'js'
+			|| languageId === 'ts'
+			|| languageId === 'vue'
+	}
+
+	export function isHTMLSyntax(languageId: AllLanguageId): boolean {
+		return languageId === 'html'
+			|| languageId === 'vue'
 	}
 
 	export function isReactScriptSyntax(languageId: AllLanguageId): boolean {