+++ server/src/languages/language-ids/types.ts
@@ -1,5 +1,5 @@
 type CSSLanguageId = 'css' | 'sass' | 'scss' | 'less'
 
-type HTMLLanguageId = 'jsx' | 'tsx' | 'js' | 'ts' | 'html'
+type HTMLLanguageId = 'jsx' | 'tsx' | 'js' | 'ts' | 'html' | 'vue'
 
 type AllLanguageId = HTMLLanguageId | CSSLanguageId