+++ server/src/languages/trees/picker.ts
@@ -11,21 +11,21 @@ export namespace Picker {
 	 * Starts with `a-z` or `_`, but can follow any non-whitespace characters.
 	 */
 	export function pickClassNames(text: string): Picked[] {
-		let re = /\w[^\s"'`}]*/g
+		let re = /(\w[^\s"'`]*)|\$\{.*?\}/g
 		let match: RegExpExecArray | null
 		let words: Picked[] = []
 
 		while (match = re.exec(text)) {
 
 			// Have interpolation characters.
-			if (/\$\{/.test(match[0])) {
+			if (!match[1]) {
 				continue
 			}
 
 			let start = match.index
 
 			words.push({
-				text: match[0],
+				text: match[1],
 				start,
 			})
 		}